+++ R/inventory.R
@@ -11,7 +11,8 @@
 #' case that tthere are no vehiles for one category of the composition, the name
 #' should be included with the number zero, for example PC = 0. The maximum
 #' number allowed is 99 per category.
-#' @param scripts Logical value for aggregate or no scripts.
+#' @param show.main Logical; Do you want to see the new main.R file?
+#' @param scripts Logical Do you want to generate or no R scripts?
 #' @param show.dir Logical value for printing the created directories.
 #' @param show.scripts Logical value for printing the created scripts.
 #' @param clear Logical value for removing recursively the directory and create
@@ -22,9 +23,7 @@
 #' sources of emissions. The structure of the directories is: daily, ef, emi,
 #' est, images, network and veh. This structure is a suggestion and the user can
 #' use another.
-#'
-#' daily: it is for storing the profiles saved as .csv files
-#'
+#''
 #' ef: it is for storing the emission factors data-frame, similar to data(fe2015)
 #' but including one column for each of the categories of the vehicular
 #' composition. For intance, if PC = 5, there should be 5 columns with emission
@@ -58,6 +57,7 @@
 #' traffic.R to generates objects of class "Vehicles".
 #' The user can rename these scripts.
 #' @export
+#' @importFrom utils file.edit
 #' @examples \dontrun{
 #' name = file.path(tempdir(), "YourCity")
 #' inventory(name = name, show.dir = TRUE, show.scripts = TRUE)
@@ -66,6 +66,7 @@
 #'
 inventory <- function(name,
                       vehcomp = c(PC = 1, LCV = 1, HGV = 1, BUS = 1, MC = 1),
+                      show.main = TRUE,
                       scripts = TRUE,
                       show.dir = TRUE,
                       show.scripts = FALSE,
@@ -294,4 +295,5 @@ inventory <- function(name,
     cat("Scripts:\n")
     print(sc)
   }
+if(show.main)  utils::file.edit(paste0(name, "/main.R"))
 }